feat: change ushort to int in histogram model

+++ deploy/IoTEventWorker/IotEventWorker/Models/Histogram.cs
@@ -1,9 +1,9 @@
 ﻿namespace IoTEventWorker.Models;
 
-public class Histogram<T>(T[] tips, int slotCount, ushort slotSecs, DateTimeOffset startTime)
+public class Histogram<T>(T[] tips, int slotCount, int slotSecs, DateTimeOffset startTime)
 {
     public T[] Tips { get; } = tips;
     public int SlotCount { get; } = slotCount;
-    public ushort SlotSecs { get; } = slotSecs;
+    public int SlotSecs { get; } = slotSecs;
     public DateTimeOffset StartTime { get; } = startTime;
 }
